Sales Fact Table
This document contains the Sales figures from January 2025 to date.

It includes information regarding the value of the sale, who closed the sale, which competitor they gained the client from and what marketing techniques were utilised to close that sale. As this information relates to the Security Industry, it also looks at when the clients alarm systems were linked to that security provider to monitor their signals, and includes their linked devices and last signal dates. 

I extracted and cleaned the data in Excel - removing any unnecessary information